+++ CHANGELOG.md
@@ -15,7 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `Open-RegistryKey`
   - `Open-RegistrySubKey`
   - `Remove-RegistrySubKey`
+- Added unit test skeletons for all public functions in the WinRegOps
 
 ### Changed
 - Added 'WisherTools.Helpers' to Nested Modules
 - Updated `build.yaml` to exclude `Modules/WisherTools.Helpers` from code coverage analysis.
+- ### Changed
+- Refactored `Open-RegistryKey` function to use new helper functions `Get-OpenBaseKey` and `Get-OpenRemoteBaseKey` to abstract static method calls for opening registry keys locally or remotely. This improves testability and modularity of the code.